Add files via upload
Adding the hand_mask2.png which is a "mask" of the right hand graphic.   The alice.txt is some phony text from Alice in Wonderland.  This is just to show proof of concept that we can make a word cloud masked to the shape of a human right-hand ---for "Data Science as a Human Right".